Refactor: Split out stream events
This splits out the streaming logic it separate files. It's a first pass to clean up the streaming logic for readability

Change-type: minor
